Move and page renders to pages folder
